+++ tidy3d/components/medium.py
@@ -884,7 +884,7 @@ class AbstractMedium(ABC, Tidy3dBaseModel):
         nonlinear_spec = values.get("nonlinear_spec")
         if val is not None and nonlinear_spec is not None:
             raise ValidationError(
-                f"For medium class {cls}, 'modulation_spec' of class {type(val)} and "
+                f"For medium class {cls.__name__}, 'modulation_spec' of class {type(val)} and "
                 f"'nonlinear_spec' of class {type(nonlinear_spec)} are "
                 "not simultaneously supported."
             )
@@ -1350,6 +1350,11 @@ class AbstractMedium(ABC, Tidy3dBaseModel):
         """Whether the medium is a PEC."""
         return False
 
+    @cached_property
+    def is_pmc(self):
+        """Whether the medium is a PMC."""
+        return False
+
     def sel_inside(self, bounds: Bound) -> AbstractMedium:
         """Return a new medium that contains the minimal amount data necessary to cover
         a spatial region defined by ``bounds``.
@@ -1740,7 +1745,7 @@ class PECMedium(AbstractMedium):
         if val is not None:
             raise ValidationError(
                 f"A 'modulation_spec' of class {type(val)} is not "
-                f"currently supported for medium class {cls}."
+                f"currently supported for medium class {cls.__name__}."
             )
         return val
 
@@ -1765,6 +1770,52 @@ class PECMedium(AbstractMedium):
 
 # PEC builtin instance
 PEC = PECMedium(name="PEC")
+
+
+# PMC keyword
+class PMCMedium(AbstractMedium):
+    """Perfect magnetic conductor class.
+
+    Note
+    ----
+
+        To avoid confusion from duplicate PMCs, must import ``tidy3d.PMC`` instance directly.
+
+
+
+    """
+
+    @pd.validator("modulation_spec", always=True)
+    def _validate_modulation_spec(cls, val):
+        """Check compatibility with modulation_spec."""
+        if val is not None:
+            raise ValidationError(
+                f"A 'modulation_spec' of class {type(val)} is not "
+                f"currently supported for medium class {cls.__name__}."
+            )
+        return val
+
+    @ensure_freq_in_range
+    def eps_model(self, frequency: float) -> complex:
+        # permittivity of a PMC.
+        return 1.0 + 0j
+
+    @cached_property
+    def n_cfl(self):
+        """This property computes the index of refraction related to CFL condition, so that
+        the FDTD with this medium is stable when the time step size that doesn't take
+        material factor into account is multiplied by ``n_cfl``.
+        """
+        return 1.0
+
+    @cached_property
+    def is_pmc(self):
+        """Whether the medium is a PMC."""
+        return True
+
+
+# PEC builtin instance
+PMC = PMCMedium(name="PMC")
 
 
 class Medium(AbstractMedium):
@@ -5643,9 +5694,10 @@ class LossyMetalMedium(Medium):
         return ax
 
 
-IsotropicUniformMediumType = Union[
+IsotropicUniformMediumFor2DType = Union[
     Medium, LossyMetalMedium, PoleResidue, Sellmeier, Lorentz, Debye, Drude, PECMedium
 ]
+IsotropicUniformMediumType = Union[IsotropicUniformMediumFor2DType, PMCMedium]
 IsotropicCustomMediumType = Union[
     CustomPoleResidue,
     CustomSellmeier,
@@ -5719,7 +5771,7 @@ class AnisotropicMedium(AbstractMedium):
         if val is not None:
             raise ValidationError(
                 f"A 'modulation_spec' of class {type(val)} is not "
-                f"currently supported for medium class {cls}. "
+                f"currently supported for medium class {cls.__name__}. "
                 "Please add modulation to each component."
             )
         return val
@@ -5852,9 +5904,18 @@ class AnisotropicMedium(AbstractMedium):
         """Whether the medium is a PEC."""
         return any(self.is_comp_pec(i) for i in range(3))
 
+    @cached_property
+    def is_pmc(self):
+        """Whether the medium is a PMC."""
+        return any(self.is_comp_pmc(i) for i in range(3))
+
     def is_comp_pec(self, comp: Axis):
         """Whether the medium is a PEC."""
         return isinstance(self.components[["xx", "yy", "zz"][comp]], PECMedium)
+
+    def is_comp_pmc(self, comp: Axis):
+        """Whether the medium is a PMC."""
+        return isinstance(self.components[["xx", "yy", "zz"][comp]], PMCMedium)
 
     def sel_inside(self, bounds: Bound):
         """Return a new medium that contains the minimal amount data necessary to cover
@@ -5945,7 +6006,7 @@ class FullyAnisotropicMedium(AbstractMedium):
         if val is not None:
             raise ValidationError(
                 f"A 'modulation_spec' of class {type(val)} is not "
-                f"currently supported for medium class {cls}."
+                f"currently supported for medium class {cls.__name__}."
             )
         return val
 
@@ -6970,6 +7031,7 @@ MediumType3D = Union[
     Medium,
     AnisotropicMedium,
     PECMedium,
+    PMCMedium,
     PoleResidue,
     Sellmeier,
     Lorentz,
@@ -7004,7 +7066,7 @@ class Medium2D(AbstractMedium):
 
     """
 
-    ss: IsotropicUniformMediumType = pd.Field(
+    ss: IsotropicUniformMediumFor2DType = pd.Field(
         ...,
         title="SS Component",
         description="Medium describing the ss-component of the diagonal permittivity tensor. "
@@ -7015,7 +7077,7 @@ class Medium2D(AbstractMedium):
         discriminator=TYPE_TAG_STR,
     )
 
-    tt: IsotropicUniformMediumType = pd.Field(
+    tt: IsotropicUniformMediumFor2DType = pd.Field(
         ...,
         title="TT Component",
         description="Medium describing the tt-component of the diagonal permittivity tensor. "
@@ -7032,7 +7094,7 @@ class Medium2D(AbstractMedium):
         if val is not None:
             raise ValidationError(
                 f"A 'modulation_spec' of class {type(val)} is not "
-                f"currently supported for medium class {cls}."
+                f"currently supported for medium class {cls.__name__}."
             )
         return val
 
@@ -7049,7 +7111,7 @@ class Medium2D(AbstractMedium):
 
     @classmethod
     def _weighted_avg(
-        cls, meds: list[IsotropicUniformMediumType], weights: list[float]
+        cls, meds: list[IsotropicUniformMediumFor2DType], weights: list[float]
     ) -> Union[PoleResidue, PECMedium]:
         """Average ``meds`` with weights ``weights``."""
         eps_inf = 1
@@ -7103,7 +7165,7 @@ class Medium2D(AbstractMedium):
             The 3D material corresponding to this 2D material.
         """
 
-        def get_component(med: MediumType3D, comp: Axis) -> IsotropicUniformMediumType:
+        def get_component(med: MediumType3D, comp: Axis) -> IsotropicUniformMediumFor2DType:
             """Extract the ``comp`` component of ``med``."""
             if isinstance(med, AnisotropicMedium):
                 dim = "xyz"[comp]
@@ -7365,7 +7427,7 @@ class Medium2D(AbstractMedium):
         return np.mean([self.ss.sigma_model(freq), self.tt.sigma_model(freq)], axis=0)
 
     @property
-    def elements(self) -> dict[str, IsotropicUniformMediumType]:
+    def elements(self) -> dict[str, IsotropicUniformMediumFor2DType]:
         """The diagonal elements of the 2D medium as a dictionary."""
         return {"ss": self.ss, "tt": self.tt}