+++ tidy3d/components/scene.py
@@ -519,9 +519,11 @@ class Scene(Tidy3dBaseModel):
 
         if isinstance(medium, MultiPhysicsMedium):
             is_pec = medium.optical is not None and medium.optical.is_pec
+            is_pmc = medium.optical is not None and medium.optical.is_pmc
             is_time_modulated = medium.optical is not None and medium.optical.is_time_modulated
         else:
             is_pec = medium.is_pec
+            is_pmc = medium.is_pmc
             is_time_modulated = medium.is_time_modulated
 
         if mat_index == 0 or medium == self.medium:
@@ -531,6 +533,11 @@ class Scene(Tidy3dBaseModel):
             # perfect electrical conductor
             plot_params = plot_params.copy(
                 update={"facecolor": "gold", "edgecolor": "k", "linewidth": 1}
+            )
+        elif is_pmc:
+            # perfect magnetic conductor
+            plot_params = plot_params.copy(
+                update={"facecolor": "purple", "edgecolor": "k", "linewidth": 1}
             )
         elif is_time_modulated:
             # time modulated medium
@@ -1284,6 +1291,11 @@ class Scene(Tidy3dBaseModel):
             # perfect electrical conductor
             plot_params = plot_params.copy(
                 update={"facecolor": "gold", "edgecolor": "k", "linewidth": 1}
+            )
+        elif medium.is_pmc:
+            # perfect magnetic conductor
+            plot_params = plot_params.copy(
+                update={"facecolor": "purple", "edgecolor": "k", "linewidth": 1}
             )
         elif isinstance(medium, Medium2D):
             # 2d material